Enhance User and Certificate handling in models and resources
Implemented several changes in the User and Certificate models and resources. Updated the User model by adding a role manager to the 'canAccessPanel' method. Added new ViewCertificate.php class in the CertificateResource. Also, in the GatewayUser model, a 'certificate' relation is now fetched by default. In GatewayUserResource and CertificateResource, form fields and table columns have been adjusted to reflect the model modifications.