Add UserObserver and tests to assign company to users
A UserObserver has been created to automatically assign the company of the authenticated user to the newly created users. This update also includes alterations in UserFactory and CompanyFactory as well as new tests to validate the functionality. Additionally, a helper function is introduced in Pest.php to create necessary roles.