Enhance Certificate handling and update Certificate Factory
The main changes focus on enhancing the certificate handling by making certain fields in the CertificateResource sortable and searchable for better user experience. Certificate Factory has been updated to include the status of the certificate request. The certificate model now has a function to generate X509 data. The "Approved" status from CertificateRequestStatus enum was removed as it seemed redundant with "Issued". Updated the ViewCertificate page to handle the approval or rejection of a certificate request.