Add certificate regeneration feature and validity selection
This update introduces the ability for certificates to be regenerated if they have been issued already. A method `canBeRegenerated()` was added to check this. Certificate validity selection feature has also been integrated which allows the selection of four different time ranges. Adjustments were made to the certificate generation process to honor the selected validity option. Changes were made to the database seeders and factories to align with these updates.